requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aiosqlite
google-api-python-client
google-auth
pydantic-settings
